fix(desktop): surface persona pack write-back failures
write_back_persona_md runs on every update_persona for pack-backed
personas and swallowed every failure behind a bare eprintln!, so a failed
write-back silently diverged the saved record from its source .persona.md
with no signal to the user.

Extract try_write_back_persona_md as a testable function that takes
resolved teams rather than an AppHandle, and return failures as
Option<String> from write_back_persona_md. update_persona forwards it as a
non-fatal writeback_warning on UpdatePersonaResult so the frontend can
surface the divergence; in-app edits still save even when the pack file is
unwritable.

Add regression tests for the legacy team shape (UUID team id, source_dir
whose basename is the team_persona_key, symlinked pack dir, persona .md
missing the runtime: key — confirmed working on current code) and for the
read-only pack file case.
